fix(harness): preserve checkpoint history exactly
Resolve Round A3 findings by enforcing byte-identical PASS prefixes and sharing authoritative Markdown visibility with raw provenance offsets.
